feat:  adjust buffer sizes to align with MATLAB logic, moved signal verification function to data_init.c
