+++ inc/data_init.h
@@ -32,4 +32,6 @@ double *get_ch_signal(double **data,
 											size_t num_cols,
 											int channel_num);
 
+int verify_signals(double *signal, size_t num_of_data, double **verify_data, size_t *verify_num_rows, size_t *verify_num_cols);
+
 #endif
